requirements.txt
langchain==0.0.208
deeplake==3.6.5
openai==0.27.8
tiktoken==0.4.0
python-dotenv
geopy
transformers
matplotlib
chromadb
streamlit
python-dotenv
streamlit-chat